backend/requirements.txt
deepagents>=0.1.0
langchain-anthropic>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
python-dotenv>=1.0.0
anyio>=4.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30.0
psycopg[binary]>=3.1
langgraph-checkpoint-postgres>=2.0.0
loguru
elasticsearch==7.10.0